refactor: 메인 페이지에 조건부 렌더링 분기 수정  및 페이지 위치 변경
